refactor(api): extract discovery handlers to request-handlers/discovery.mjs (#510)
First module of the workers/api.mjs de-monolith (Wave B maintainability). Pure
code-motion, zero behavior change: moves the machine-discovery surfaces —
homepage, /.well-known/api-catalog, MCP server-card, agent-tools specs, and the
per-netuid health badge SVG (+ the BADGE_*/DISCOVERY_* constants and badge/SVG
helpers) — into a new workers/request-handlers/discovery.mjs. api.mjs imports
the five handlers + BADGE_SVG_PATTERN for dispatch; now-unused imports
(listToolDefinitions, feedLinkHeader, the agent-tool-spec builders, etc.) are
dropped. No logic edits, no test changes (discovery fns aren't imported by
tests). api.mjs 3517 -> 3214 LOC.

Verified: npm test 46 files / 1117 tests green (matches pristine main), eslint +
prettier clean. Next modules: shared (analytics/slug/network helpers), then
rpc/health/rest/webhooks.
